Replaces how it works with event badge info, few content tweaks
This replaces the "how it works" section with the event badging info (instead of the project badging info) and makes a few other content tweaks.

There are a lot of changes in this PR and it's hard to envision if it will display properly. Sorry if there are mistakes in here! 💖